Start debugging intra-doc link bug
The current fix is to just make sure the base namespace is exposed with
`use`. But it seems like this is a bug with rustdoc itself. So I also
want to debug rustdoc.
Debugging default_plugins.rs is the simplest thing to debug because it
is processed first before other modules.